Create a function to get 'SelectControl' options
By default the latest page was being selected in the 'SelectControl' which
is not a good user experience. So created a function that adds the
default option first.

Also selecting this default option causes the JSON parser to fail as it
has 'null' value. So had to set it to custom JSON value.